backend/requirements.txt
annotated-types==0.6.0
appnope==0.1.3
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Flask==3.0.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
inflect==7.0.0
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
six==1.16.0
sqlacodegen-v2==0.1.4
SQLAlchemy==2.0.23
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.12
Werkzeug==3.0.1
